Adding check for nil fields (#1)
I ran into a situation where a bot has the POST request stored somewhere prior to me adding the honeypot to the form.
This way it is bypassing the filter since it doesn’t have the honeypot field filled at all in the request while all the others are filled, making it look legitimate.
I added a check to see if the honeypot field is nil. This is because when a user submits the form from his browser, the honeypot field will be an empty string and not nil.